requirements.txt
pytest~=8.4.1
requests~=2.32.4
allure-python-commons~=2.14.3